TeamEzcgp/ResearchAugmentor/requirements.txt
Augmentor==0.2.8
cycler==0.10.0
future==0.18.2
kiwisolver==1.1.0
matplotlib==3.1.3
numpy==1.18.1
Pillow==7.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
six==1.14.0
tqdm==4.42.1
